Code cleaning and optimization in users
