+++ app/admin/users/UsersTable/index.tsx
@@ -1,17 +1,15 @@
 "use client";
 
-import { useUsers } from "@/hooks/useUsers";
+import { useMemo } from "react";
+
+import { useGetUsers } from "@/hooks/users/useGetUsers";
 
 import { UsersTableUser } from "@/interfaces/UsersTableUser";
 
 import { CustomTable } from "@/components/Table";
 
 export const UsersTable = () => {
-  const users = useUsers();
-
-  if (users.isLoading) {
-    return <div>Loading</div>
-  }
+  const users = useGetUsers();
 
   const tableHead = [
     "Username",
@@ -23,15 +21,21 @@ export const UsersTable = () => {
     "Updated at",
   ];
 
-  const tableBody = users?.data?.map(({id, username, email, emailVerified, role, active, createdAt, updatedAt}: UsersTableUser) => [
-    username,
-    email,
-    emailVerified ? 'Yes' : 'No',
-    role,
-    active ? 'Yes' : 'No',
-    createdAt,
-    updatedAt,
-  ]);
+  const tableBody = useMemo(() => {
+    return users?.data?.map(({id, username, email, emailVerified, role, active, createdAt, updatedAt}: UsersTableUser) => [
+      username,
+      email,
+      emailVerified ? 'Yes' : 'No',
+      role,
+      active ? 'Yes' : 'No',
+      createdAt,
+      updatedAt,
+    ]);
+  }, [users]);
+
+  if (users.isLoading) {
+    return <div>Loading users, please wait...</div>
+  }
 
   return (
     <CustomTable
